agribot_backend/requirements.txt
gradio>=4.0.0
transformers>=4.35.0
torch
deep-translator
gtts
openai-whisper
sentencepiece
